Adicionando model e controller para taxonomia, adicionando métodos respectivos e ajustando nomeclatura de arquivos/funções
